Rename input to raw_input
